Gates and Policies

+++ app/Http/Controllers/Api/EventController.php
@@ -2,13 +2,18 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use App\Http\Resources\EventResource;
-use App\Http\Traits\CanLoadRelationships;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Resources\EventResource;
+use App\Http\Traits\CanLoadRelationships;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
 
-class EventController extends Controller
+class EventController extends Controller implements HasMiddleware
 {
     use CanLoadRelationships;
 
@@ -18,14 +23,25 @@ class EventController extends Controller
         'attendees.user'
     ];
 
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     public function index()
     {
+        Gate::authorize('viewAny', Event::class);
+
         $query = $this->loadRelationships(Event::query());
         return EventResource::collection($query->latest()->paginate());
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create', Event::class);
+
         $validated = $request->validate([
             'name' => 'required|string|min:5|max:255',
             'description' => 'nullable|string',
@@ -35,25 +51,21 @@ class EventController extends Controller
 
         $event = Event::create([
             ...$validated,
-            'user_id' => 1
+            'user_id' => $request->user()->id
         ]);
 
         return new EventResource($this->loadRelationships($event));
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Event $event)
     {
         return new EventResource($this->loadRelationships($event));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Event $event)
     {
+        Gate::authorize('update', $event);
+
         $validated = $request->validate([
             'name' => 'sometimes|string|min:5|max:255',
             'description' => 'nullable|string',
@@ -61,18 +73,17 @@ class EventController extends Controller
             'end_time' => 'sometimes|date|after:start_time',
         ]);
 
-        $updated = $event->update($validated); // bool - check if update worked
+        $event->update($validated);
 
         return new EventResource($this->loadRelationships($event));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Event $event)
     {
+        Gate::authorize('delete', $event);
+
         $event->delete();
 
-        return response(status: 204); // no response
+        return response(status: 204);
     }
 }
